fix: accumulate fractional MC-truth votes as double in Tracklet::link2MCTrack
The per-track vote weight 1./nTot was accumulated into a
std::unordered_map<int, int>, truncating each increment to 0, so the
'>min fraction' threshold never fired and the function returned the last
point's track ID rather than the dominant track. Use a double-valued map.

+++ strawtubes/Tracklet.cxx
@@ -39,7 +39,7 @@ Tracklet::~Tracklet() = default;
 Int_t Tracklet::link2MCTrack(std::vector<strawtubesPoint>* strawPoints,
                              Float_t min) {
   Int_t nTot = aTracklet.size();
-  std::unordered_map<int, int> MC;
+  std::unordered_map<int, double> MC;
   Int_t trackID = -1;
   for (std::vector<int>::size_type i = 0; i != aTracklet.size(); i++) {
     trackID = (*strawPoints)[aTracklet[i]].GetTrackID();
